+++ java/sqlbuilder/sqlbuilder-impl/src/main/java/com/provys/sqlbuilder/impl/WhereCondSimple.java
@@ -36,7 +36,7 @@ public class WhereCondSimple implements WhereCond {
     }
 
     @Override
-    public int getCost() {
+    public double getCost() {
         return 1000;
     }
